pyproject.toml
[tool.poetry]
name = "take-home-test"
version = "0.1.0"
description = "Sertis data engineering take home test"
authors = ["J. Doe"]

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "3.1.2"
python-dotenv = "1.0.0"
psycopg2 = "2.9.6"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
